backend/requirements.txt
django==3.2.25
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==2.4.0
daphne==2.5.0
opencv-contrib-python==4.6.0.66
numpy==1.24.3
Pillow==8.4.0
dlib==19.24.0
